% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_alignment.R
\name{add_timestamp}
\alias{add_timestamp}
\title{Get the timestamp for a row of adv data}
\usage{
add_timestamp(adv_data, stat_df, i, interval = 0.0625)
}
\arguments{
\item{adv_data}{an adv dataframe}

\item{stat_df}{corresponding status dataframe}

\item{i}{row index}

\item{interval}{seconds between data lines}
}
\value{
timestamp of the current d line
}
\description{
Get the timestamp for a row of adv data
}
