% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_alignment.R
\name{add_timestamps}
\alias{add_timestamps}
\title{Add timestamps to ADV data}
\usage{
add_timestamps(adv_data, stat_df)
}
\arguments{
\item{adv_data}{adv dataframe}

\item{stat_df}{adv status dataframe}
}
\value{
the adv dataframe with timestamps added
}
\description{
Add timestamps to ADV data
}
