% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{cal_ph}
\alias{cal_ph}
\title{calibrate pH with fit parameters}
\usage{
cal_ph(ph_counts, temp, ph_fit)
}
\arguments{
\item{ph_counts}{A vector of ph counts}

\item{temp}{A vector of temps}

\item{ph_fit}{A named vector of lm coefficents (int, ph, temp)}
}
\value{
A vector of calibrated pH
}
\description{
calibrate pH with fit parameters
}
