% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{cal_temp}
\alias{cal_temp}
\title{Calibrate Rinko temp}
\usage{
cal_temp(raw_temp, rinko_cals)
}
\arguments{
\item{raw_temp}{A vector of raw temp voltage measurements}

\item{rinko_cals}{A named vector of rinko calibration coefficents}
}
\value{
A vector of calibrated temperatures
}
\description{
Calibrate Rinko temp
}
