% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{calc_flux}
\alias{calc_flux}
\title{Calculate fluxes by eddy covariance method}
\usage{
calc_flux(
  quantity,
  velocity,
  time,
  frequency,
  low_cutoff = 1/(15 * 60),
  high_cutoff = 0.125
)
}
\arguments{
\item{quantity}{measured quantity}

\item{velocity}{measured velocity}

\item{time}{timestamps for quantity and velocity}

\item{frequency}{sampling or resampled data frequency}

\item{low_cutoff}{low cutoff in Hz}

\item{high_cutoff}{high cutoff in Hz}
}
\value{
the flux in units of \code{quantity} per m^2 per length of \code{time}
}
\description{
Calculates flux for the unit of data provided, typically hourly.
Run over data grouped by time period to get flux over time.
}
