% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{calc_response}
\alias{calc_response}
\title{Calculate mean response (spectral power) in a frequency range}
\usage{
calc_response(quantity, time, frequency, freq_range = c(0.1, 1))
}
\arguments{
\item{quantity}{Vector of sensor data}

\item{time}{Vector of corresponding timestamps}

\item{frequency}{Frequency of measurement}

\item{freq_range}{Range of frequencies to calc power average}
}
\value{
the mean value
}
\description{
Calculate mean response (spectral power) in a frequency range
}
