% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_alignment.R
\name{calc_ts}
\alias{calc_ts}
\title{Calculate D line timestamp from S line timestamps}
\usage{
calc_ts(cur_count, prev_ts, next_ts, prev_count, next_count, interval = 0.0625)
}
\arguments{
\item{cur_count}{Current D line count}

\item{prev_ts}{Last S line timestamp}

\item{next_ts}{Next S line timestamp}

\item{prev_count}{Count at last S line timestamp}

\item{next_count}{Count at next S line timestamp}

\item{interval}{Time between D lines}
}
\value{
timestamp of current D line
}
\description{
Calculate D line timestamp from S line timestamps
}
