% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_alignment.R
\name{correct_status_timestamp_jitter}
\alias{correct_status_timestamp_jitter}
\title{Correct adv timestamp using jitter correction}
\usage{
correct_status_timestamp_jitter(timestamp, adv_timestamp)
}
\arguments{
\item{timestamp}{A vector of teensy lander timestamps}

\item{adv_timestamp}{A vector of ADV timestamps}
}
\value{
A vector of corrected timestamps.
}
\description{
Provides a per-file timestamp that does not have duplicated
or missing timestamps due to serial packet arrival time jitter.
}
\details{
Removes future timestamps to prevent "2038" problem.
}
