% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{count_missing}
\alias{count_missing}
\title{Count missing elements in repeating 0-254 sequence}
\usage{
count_missing(numlist)
}
\arguments{
\item{numlist}{a vector of integers}
}
\value{
the number of missing elements
}
\description{
Count missing elements in repeating 0-254 sequence
}
