% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_alignment.R
\name{count_tzero}
\alias{count_tzero}
\title{find count at status line index}
\usage{
count_tzero(data_df, status_row)
}
\arguments{
\item{data_df}{adv data dataframe}

\item{status_row}{row_num of status row}
}
\value{
count
}
\description{
Assume first line after status is same time as status
}
