% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_alignment.R
\name{counts_between}
\alias{counts_between}
\title{Difference between two numbers in a circular 1 byte number}
\usage{
counts_between(c1, c2)
}
\arguments{
\item{c1}{first number}

\item{c2}{second number}
}
\value{
difference
}
\description{
Difference between two numbers in a circular 1 byte number
}
