% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{fit_ph_model}
\alias{fit_ph_model}
\title{calibrate pH with fit parameters}
\usage{
fit_ph_model(ph_counts, temp, ph_fit)
}
\arguments{
\item{ph_counts}{A vector of ph counts}

\item{temp}{A vector of temps}

\item{ph_fit}{A named vector of lm coeffients}
}
\value{
A vector of calibrated pH
}
\description{
calibrate pH with fit parameters
}
