% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fix_timestamp_jitter}
\alias{fix_timestamp_jitter}
\title{Fix status timestamp jitter}
\usage{
fix_timestamp_jitter(timestamp, adv_timestamp)
}
\arguments{
\item{timestamp}{A vector of status timestamps}

\item{adv_timestamp}{A vector of adv status timestamps}
}
\value{
The corrected timestamp vector
}
\description{
Looks for timestamps same as previous timestamp.
Adds 1s to duplicate timestamp
If there is 1s gap after previous timestamp,
and gap is not in ADV timestamp, remove gap.
}
