% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{fraction_missing}
\alias{fraction_missing}
\title{Calculate Fraction of missing elements}
\usage{
fraction_missing(numlist)
}
\arguments{
\item{numlist}{a vector of integers}
}
\value{
a fraction of missing elements
}
\description{
Calculate Fraction of missing elements
}
