% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{gems_add_metadata}
\alias{gems_add_metadata}
\title{Parse GEMS metadata}
\usage{
gems_add_metadata(df_raw)
}
\arguments{
\item{df_raw}{a raw GEMS dataframe}
}
\value{
a dataframe with added metadata
}
\description{
Add row numbers, line type, number and line of send (web data),
}
