% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{gems_clean_rga}
\alias{gems_clean_rga}
\title{clean RGA data}
\usage{
gems_clean_rga(rga)
}
\arguments{
\item{rga}{}
}
\value{
cleaned RGA data
}
\description{
Final output columns are selected in gems_parse_file
}
