% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{gems_clean_status}
\alias{gems_clean_status}
\title{Clean status data}
\usage{
gems_clean_status(status)
}
\arguments{
\item{status}{}
}
\description{
Final output columns are selected in gems_parse_file
}
