% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{gems_missing}
\alias{gems_missing}
\title{Calculate number of missing lines}
\usage{
gems_missing(count, line = NULL)
}
\arguments{
\item{count}{Vector of ADV count}

\item{line}{vector indicating line of send}
}
\value{
A vector with a guess at the number of missing lines from
previous to current
}
\description{
Calculate number of missing lines
}
