% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{gems_parse_file}
\alias{gems_parse_file}
\title{Read gems data from file and parse into dataframes}
\usage{
gems_parse_file(file, clean = FALSE)
}
\arguments{
\item{file}{A file path in gems data format}

\item{clean}{Filter bad data and timestamps if true}
}
\value{
a list containing gems post_times, rga data, status data, and ADV data
}
\description{
Runs file parsers, time alignment, and selects columns for output.
}
