% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{gems_parse_files}
\alias{gems_parse_files}
\title{Parallelized Read gems data from files and parse into dataframes}
\usage{
gems_parse_files(files, clean = FALSE)
}
\arguments{
\item{files}{A list of file paths containing gems data}
}
\value{
a list containing rga data, status data, and ADV data
}
\description{
Parallelized Read gems data from files and parse into dataframes
}
