% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{gems_post_times}
\alias{gems_post_times}
\title{parse GEMS post times}
\usage{
gems_post_times(df)
}
\arguments{
\item{df}{a GEMS dataframe with added metadata}
}
\value{
a dataframe of post times
}
\description{
parse GEMS post times
}
