% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{gems_read_file}
\alias{gems_read_file}
\title{GEMS raw file parser}
\usage{
gems_read_file(file)
}
\arguments{
\item{files}{a vector of file paths}
}
\value{
a dataframe with row_num, type, and unparsed data
}
\description{
Parse raw files to a df with row_num and type
Remove type chars from beginning of lines
}
