% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{gems_rga_data}
\alias{gems_rga_data}
\title{parse RGA data}
\usage{
gems_rga_data(df)
}
\arguments{
\item{df}{a dataframe with added metadata}
}
\value{
a dataframe of RGA data
}
\description{
parse RGA data
}
