% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{gems_status_data}
\alias{gems_status_data}
\title{parse GEMS status data}
\usage{
gems_status_data(df)
}
\arguments{
\item{df}{a GEMS dataframe with added metadata}
}
\value{
a dataframe of status data
}
\description{
parse GEMS status data
}
