% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{gems_temp_data}
\alias{gems_temp_data}
\title{parse surface teensy temp}
\usage{
gems_temp_data(df)
}
\arguments{
\item{df}{a dataframe with added metadata}
}
\value{
a dataframe of surface teensy temp data
}
\description{
parse surface teensy temp
}
