% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{gems_turbo_status}
\alias{gems_turbo_status}
\title{parse GEMS turbo status data}
\usage{
gems_turbo_status(df)
}
\arguments{
\item{df}{a GEMS dataframe with added metadata}
}
\value{
a dataframe of status data
}
\description{
parse GEMS turbo status data
}
