% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{generate_o2_model}
\alias{generate_o2_model}
\title{Fit a O2 model to the given data and return coefficients}
\usage{
generate_o2_model(seaphox_data, lecs_data)
}
\arguments{
\item{seaphox_data}{a seaphox dataframe}

\item{adv_data}{an adv dataframe}
}
\value{
A named vector of lm coefficents
}
\description{
TODO: work with o2 concentration instead of saturation
}
