% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{generate_ph_model}
\alias{generate_ph_model}
\title{Fit a ph model to the given data and return coefficients}
\usage{
generate_ph_model(seaphox_data, lecs_data)
}
\arguments{
\item{seaphox_data}{a seaphox dataframe}

\item{adv_data}{an adv dataframe}
}
\value{
A named vector of lm coefficents
}
\description{
Fit a ph model to the given data and return coefficients
}
