% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{hourly_flux}
\alias{hourly_flux}
\title{Calculate hourly fluxes}
\usage{
hourly_flux(adv_data, frequency)
}
\arguments{
\item{adv_data}{Processed adv_data with temp, O2, and H+ concentrations}

\item{frequency}{Frequency of data (after resampling)}
}
\value{
A dataframe of hourly fluxes
}
\description{
Calculate hourly fluxes
}
