% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{lecs_adv_data}
\alias{lecs_adv_data}
\title{parse LECS ADV data}
\usage{
lecs_adv_data(df)
}
\arguments{
\item{df}{a LECS dataframe with added metadata}
}
\value{
a dataframe of ADV data
}
\description{
Run time alignment after this
Final output columns are selected in lecs_parse_file
}
