% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{lecs_clean_status}
\alias{lecs_clean_status}
\title{Clean status data}
\usage{
lecs_clean_status(status)
}
\arguments{
\item{status}{}
}
\description{
Final output columns are selected in lecs_parse_file
}
