% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{lecs_create_db}
\alias{lecs_create_db}
\title{Initialize a DuckDB for storing LECS data}
\usage{
lecs_create_db(lecs_data, db_file)
}
\arguments{
\item{lecs_data}{}
}
\value{
a db connection
}
\description{
Initialize a DuckDB for storing LECS data
}
