% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers_new.R
\name{lecs_met_data}
\alias{lecs_met_data}
\title{parse LECS met data}
\usage{
lecs_met_data(df)
}
\arguments{
\item{df}{a LECS dataframe with added metadata}
}
\value{
a dataframe of met data
}
\description{
parse LECS met data
}
