% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{lecs_parse_file_db}
\alias{lecs_parse_file_db}
\title{Read LECS data from file and parse into DuckDB}
\usage{
lecs_parse_file_db(con, file, clean = FALSE)
}
\arguments{
\item{file}{A file path in LECS data format}

\item{clean}{Filter bad data and timestamps if true}
}
\value{
none
}
\description{
Read LECS data from file and parse into DuckDB
}
