% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{lecs_parse_files}
\alias{lecs_parse_files}
\title{Read LECS data from files and parse into dataframes}
\usage{
lecs_parse_files(files)
}
\arguments{
\item{files}{A list of file paths containing LECS data}
}
\value{
a list containing met data, status data, and ADV data
}
\description{
Read LECS data from files and parse into dataframes
}
