% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{lecs_parse_files_db}
\alias{lecs_parse_files_db}
\title{Read LECS data from files and parse into duckdb}
\usage{
lecs_parse_files_db(con, files, clean = FALSE)
}
\arguments{
\item{files}{A list of file paths containing LECS data}
}
\value{
a list containing met data, status data, and ADV data
}
\description{
Read LECS data from files and parse into duckdb
}
