% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{lecs_parse_files_p}
\alias{lecs_parse_files_p}
\title{Parallelized Read LECS data from files and parse into dataframes}
\usage{
lecs_parse_files_p(files, clean = FALSE)
}
\arguments{
\item{files}{A list of file paths containing LECS data}
}
\value{
a list containing met data, status data, and ADV data
}
\description{
Parallelized Read LECS data from files and parse into dataframes
}
