% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers_new.R
\name{lecs_post_times}
\alias{lecs_post_times}
\title{parse LECS post times}
\usage{
lecs_post_times(df)
}
\arguments{
\item{df}{a LECS dataframe with added metadata}
}
\value{
a dataframe of post times
}
\description{
parse LECS post times
}
