% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{lecs_process_data}
\alias{lecs_process_data}
\title{Process all LECS files in a dated folder}
\usage{
lecs_process_data(
  date,
  file_dir = NULL,
  out_dir = NULL,
  dedupe = FALSE,
  resample = FALSE
)
}
\arguments{
\item{date}{A date to use for constructing output filenames and default file path}

\item{file_dir}{An optional directory to use instead of the default}

\item{out_dir}{An optional directory to use instead of the default}

\item{dedupe}{Set TRUE to remove lines with duplicate timestamps}

\item{resample}{Set to "second", "minute", etc. to output downsampled data}
}
\description{
Process all LECS files in a dated folder
}
