% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{lecs_process_data_db}
\alias{lecs_process_data_db}
\title{Process all LECS files}
\usage{
lecs_process_data_db(
  date = NULL,
  file_dir = NULL,
  files = NULL,
  out_dir = "",
  clean = TRUE
)
}
\arguments{
\item{date}{A date to use for constructing output filenames and default file path}

\item{file_dir}{An optional directory to use instead of the default}

\item{files}{An optional list of files to process. Supersedes \code{date} and
\code{file_dir} for generating file list}

\item{out_dir}{An optional directory to use instead of the default}

\item{clean}{Set TRUE to remove bad data and timestamps}
}
\description{
Uses DuckDB to handle larger-than-memory data, and stores data in
Arrow/Parquet format.
}
