% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{lecs_read_parse_web}
\alias{lecs_read_parse_web}
\title{Read LECS data from web and parse into dataframes}
\usage{
lecs_read_parse_web(start_date = NULL)
}
\arguments{
\item{start_date}{Date to get date from. Date object or string coerceable
to a date}
}
\value{
a list containing LECS post_times, met data, status data, and ADV data
}
\description{
Read LECS data from web and parse into dataframes
}
