% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers_new.R
\name{lecs_status_data}
\alias{lecs_status_data}
\title{parse LECS status data}
\usage{
lecs_status_data(df)
}
\arguments{
\item{df}{a LECS dataframe with added metadata}
}
\value{
a dataframe of status data
}
\description{
parse LECS status data
}
