% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_alignment.R
\name{make_gems_ts}
\alias{make_gems_ts}
\title{Apply timestamps to GEMS ADV data}
\usage{
make_gems_ts(adv_data, status)
}
\arguments{
\item{adv_data}{A dataframe with ADV data lines}

\item{status}{A dataframe with parsed timestamps and original row_nums}
}
\value{
ADV data with timestamps added
}
\description{
Version from R2evans on stack overflow
}
