% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_alignment.R
\name{make_lecs_ts}
\alias{make_lecs_ts}
\title{Apply timestamps to LECS ADV data}
\usage{
make_lecs_ts(df, status, adv_data)
}
\arguments{
\item{df}{A dataframe with row_num and type from the original file}

\item{status}{A dataframe with parsed timestamps and original row_nums}

\item{adv_data}{A dataframe with ADV data lines}
}
\value{
ADV data with timestamps added
}
\description{
Dplyr version from June Choe (github.com/yjunechoe)
}
