% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_alignment.R
\name{next_status}
\alias{next_status}
\title{Determine next status line index}
\usage{
next_status(stat_df, row_num)
}
\arguments{
\item{stat_df}{Status dataframe}

\item{row_num}{row_num of data line}
}
\value{
row_num of next status line
}
\description{
Determine next status line index
}
