% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{o2_ml_l_to_umol_l}
\alias{o2_ml_l_to_umol_l}
\title{Convert Oxygen Concentration from ml/L to μmol/L}
\usage{
o2_ml_l_to_umol_l(oxygen_ml_L, temperature_celsius)
}
\arguments{
\item{oxygen_ml_L}{Numeric. Dissolved oxygen concentration in ml/L.}

\item{temperature_celsius}{Numeric. Water temperature in degrees Celsius.}
}
\value{
Numeric. Dissolved oxygen concentration in μmol/L.
}
\description{
This function converts dissolved oxygen concentration from milliliters per liter (ml/L)
to micromoles per liter (μmol/L). It accounts for temperature effects on the molar volume of oxygen.
}
\examples{
# Example parameters
oxygen_ml_L <- 5.0
temperature_celsius <- 25.0

# Convert oxygen concentration
o2_ml_l_to_umol_l(oxygen_ml_L, temperature_celsius)

}
