% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{o2_sat_to_conc}
\alias{o2_sat_to_conc}
\title{Calculate O2 concentration}
\usage{
o2_sat_to_conc(
  do_percent,
  temp,
  salinity = 31.26,
  sea_pressure = 0.2,
  lon = -70.700833,
  lat = 41.516944
)
}
\arguments{
\item{do_percent}{Calibrated DO saturation percent}

\item{temp}{Conservative temp in deg C}

\item{salinity}{Absolute salinity in g/kg}

\item{sea_pressure}{Absolute pressure - 10.1325 dbar}

\item{lon}{longitude in decimal degrees}

\item{lat}{latitude in decimal degrees}
}
\value{
O2 concentration in umol/kg
}
\description{
Uses GSW. Assume sal, pressure, lon, lat if not provided.
}
