% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_met}
\alias{parse_met}
\title{Parse met Data from MCU SD card files}
\usage{
parse_met(files)
}
\arguments{
\item{files}{a list of data files}
}
\value{
A data frame of met data
}
\description{
Parse met Data from MCU SD card files
}
