% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_status}
\alias{parse_status}
\title{Parse status Data from MCU SD card files}
\usage{
parse_status(files)
}
\arguments{
\item{files}{a list of data files}
}
\value{
A data frame of ADV status data
}
\description{
Parse status Data from MCU SD card files
}
