% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_statusmsg}
\alias{parse_statusmsg}
\title{Parse status messages from MCU SD card files}
\usage{
parse_statusmsg(files)
}
\arguments{
\item{files}{a list of data files}
}
\value{
A data frame of status messages
}
\description{
Parse status messages from MCU SD card files
}
