% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_web}
\alias{parse_web}
\title{Parse data from LECS website}
\usage{
parse_web(file)
}
\arguments{
\item{file}{}
}
\value{
a dataframe in LECS SD file format
}
\description{
Currently need to copy data from https://gems.whoi.edu/LECSrawdata/
and paste to a file. Aiming to make an API eventually!
}
