% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{pH_to_conc}
\alias{pH_to_conc}
\title{Calculate H+ concentration}
\usage{
pH_to_conc(ph)
}
\arguments{
\item{ph}{Calibrated pH}
}
\value{
H+ concentration in mol/L
}
\description{
Calculate H+ concentration
}
