% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mean}
\alias{plot_mean}
\title{Plot timeseries pointrange with mean and error}
\usage{
plot_mean(data, parameter, err)
}
\arguments{
\item{data}{A dataframe with \code{timestamp_mean}}

\item{parameter}{Parameter to plot.}

\item{err}{Error in parameter}
}
\value{
A ggplot object
}
\description{
Plot timeseries pointrange with mean and error
}
