% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ts}
\alias{plot_ts}
\title{Timeseries plot}
\usage{
plot_ts(data, field, timestamp = "timestamp", title = field)
}
\arguments{
\item{data}{A dataframe}

\item{field}{Character string of name of column.}

\item{timestamp}{Character string of name timestamp column.}

\item{title}{Character string of plot title.}
}
\value{
Timeseries plot
}
\description{
Converts the given parameter to a timeseries
and plots with \code{dygraphs}
}
