% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_alignment.R
\name{prev_status}
\alias{prev_status}
\title{Determine previous status line index}
\usage{
prev_status(stat_df, row_num)
}
\arguments{
\item{stat_df}{Status dataframe}

\item{row_num}{row_num of data line}
}
\value{
row_num of previous status line
}
\description{
Determine previous status line index
}
