% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{read_adv_status}
\alias{read_adv_status}
\title{Parse ADV status file}
\usage{
read_adv_status(file, ts_offset = 0)
}
\arguments{
\item{file}{ADV .sen status file}

\item{ts_offset}{time to add to correct timestamp in seconds}
}
\value{
A data frame of status data
}
\description{
Parse ADV status file
}
