% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{read_lecs_web}
\alias{read_lecs_web}
\title{Read LECS data from website}
\usage{
read_lecs_web()
}
\value{
a single column data frame
}
\description{
Reads HTML table from LECSrawdata into a single column data frame
}
