% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{read_par_odyssey}
\alias{read_par_odyssey}
\title{Read Underwater PAR Odyssey files}
\usage{
read_par_odyssey(file)
}
\arguments{
\item{file}{}
}
\value{
a data frame of PAR data
}
\description{
Read Underwater PAR Odyssey files
}
