% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{read_prooceanus}
\alias{read_prooceanus}
\title{Read ProOceanus CO2 files}
\usage{
read_prooceanus(file)
}
\arguments{
\item{file}{}
}
\value{
a data frame of CO2 data
}
\description{
Read ProOceanus CO2 files
}
