% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{read_seaphox}
\alias{read_seaphox}
\title{Read seaphox data}
\usage{
read_seaphox(file)
}
\arguments{
\item{file}{seaphox file path}
}
\value{
a dataframe of seaphox data
}
\description{
Read seaphox data
}
