% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{read_statusmsg}
\alias{read_statusmsg}
\title{Read status messages from given file}
\usage{
read_statusmsg(file)
}
\arguments{
\item{file}{file to read}
}
\description{
Read status messages from given file
}
