% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{lecs_adv_data}
\alias{lecs_adv_data}
\title{parse LECS ADV data}
\usage{
lecs_adv_data(df, rinko_cals)
}
\arguments{
\item{df}{a LECS dataframe with added metadata}

\item{rinko_cals}{a list of rinko cal factors for temp and O2}
}
\value{
a dataframe of ADV data
}
\description{
Run time alignment after this
}
