% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_data}
\alias{parse_data}
\title{Parse ADV Data from MCU SD card files}
\usage{
parse_data(files)
}
\arguments{
\item{files}{a list of data files}
}
\value{
A data frame of ADV data
}
\description{
Parse ADV Data from MCU SD card files
}
