% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{read_lecs_web}
\alias{read_lecs_web}
\title{Read LECS data from website}
\usage{
read_lecs_web(start_date = NULL)
}
\arguments{
\item{start_date}{Date to get date from. Date object or string coerceable
to a date}
}
\value{
a single column data frame
}
\description{
Reads HTML table from LECS_data into a single column data frame
}
